TG-04: corrections of PR
