+++ package.json
@@ -26,8 +26,7 @@
   },
   "dependencies": {
     "express": "^4.18.0",
-    "nodemon": "^2.0.15",
-    "uuid": "^8.3.2"
+    "nodemon": "^2.0.15"
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
